renaming package and adding unittest
