+++ apps/apiman-manager-api/src/main/java/io/fabric8/apiman/ManagerApiMicroServiceApplication.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.apiman.manager.api.war.micro;
+package io.fabric8.apiman;
 
 import io.apiman.manager.api.rest.impl.ActionResourceImpl;
 import io.apiman.manager.api.rest.impl.ApiManagerApplication;
